Implement change meter values in HomeScreenViewModel

+++ app/src/main/java/com/serhiiromanchuk/utilitybills/presentation/viewmodel/InsertUtilityServiceViewModel.kt
@@ -65,7 +65,7 @@ class InsertUtilityServiceViewModel @Inject constructor(
                         name = name,
                         tariff = tariff.toDouble(),
                         isMeterAvailable = isMeterAvailable,
-                        previousValue = previousValue.toInt(),
+                        previousValue = previousValue,
                         unitOfMeasurement = unitOfMeasurement
                     )
                 )
